requirements.txt
requests
PyPDF2
python-docx
googlesearch-python
openai
 